+++ src/pl/tlewandster/task5/Product.java
@@ -6,9 +6,9 @@ public class Product {
     private String name;
     private BigDecimal price;
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, Double price) {
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 
     public void showProduct() {
